Add report to generated -o argument.
