+++ packages/cli/src/index.ts
@@ -19,10 +19,9 @@
  */
 import { Command } from 'commander';
 import events from 'events';
-import { existsSync, writeFile, ensureDir } from 'fs-extra';
+import { existsSync } from 'fs-extra';
 // @ts-ignore Package does not support typescript.
 import Spinnies from 'spinnies';
-import fs from 'fs';
 import path from 'path';
 import { I18n } from '@google-psat/i18n';
 import {
@@ -39,7 +38,6 @@ import {
   LIBRARIES,
   detectMatchingSignatures,
 } from '@google-psat/library-detection';
-import URL from 'node:url';
 
 /**
  * Internal dependencies.
@@ -47,7 +45,7 @@ import URL from 'node:url';
 import {
   fetchDictionary,
   getUrlListFromArgs,
-  saveCSVReports,
+  saveReports,
   askUserInput,
   generatePrefix,
   localeValidator,
@@ -55,13 +53,12 @@ import {
   filePathValidator,
   urlValidator,
   numericValidator,
-  getOutputFilePath,
 } from './utils';
 import { redLogger } from './utils/coloredLoggers';
+import saveResultsAsHTML from './utils/saveResultAsHTML';
 
 events.EventEmitter.defaultMaxListeners = 15;
 
-const isProduction = process.env.NODE_ENV === 'production';
 const program = new Command();
 
 const isFromNPMRegistry = !existsSync(
@@ -144,76 +141,6 @@ program
 
 program.parse();
 
-const saveResultsAsHTML = async (
-  outDir: string,
-  result: CompleteJson | CompleteJson[],
-  isSiteMap: boolean
-) => {
-  let htmlText = '';
-
-  await ensureDir(outDir);
-
-  if (
-    existsSync(
-      path.resolve(
-        __dirname +
-          '../../node_modules/@google-psat/cli-dashboard/dist/index.html'
-      )
-    )
-  ) {
-    htmlText = fs.readFileSync(
-      path.resolve(
-        __dirname +
-          '../../node_modules/@google-psat/cli-dashboard/dist/index.html'
-      ),
-      'utf-8'
-    );
-
-    if (!isProduction) {
-      fs.copyFileSync(
-        path.resolve(
-          __dirname +
-            '../../node_modules/@google-psat/cli-dashboard/dist/index.js'
-        ),
-        outDir + '/index.js'
-      );
-    }
-  } else {
-    htmlText = fs.readFileSync(
-      path.resolve(__dirname + '../../../cli-dashboard/dist/index.html'),
-      'utf-8'
-    );
-
-    if (!isProduction) {
-      fs.copyFileSync(
-        path.resolve(__dirname + '../../../cli-dashboard/dist/index.js'),
-        outDir + '/index.js'
-      );
-    }
-  }
-
-  const messages = I18n.getMessages();
-
-  const html =
-    htmlText.substring(0, htmlText.indexOf('</head>')) +
-    `<script id='JSONDATASCRIPT'>
-    window.PSAT_DATA = ${JSON.stringify({
-      json: result,
-      type: isSiteMap ? 'sitemap' : 'url',
-      selectedSite: outDir?.trim()?.slice(6) ?? '',
-      translations: messages,
-    })}</script>` +
-    htmlText.substring(htmlText.indexOf('</head>'));
-
-  const outputFilePath = getOutputFilePath(outDir);
-  const outFileFullDir = path.resolve(outputFilePath);
-  const htmlBlob = new Blob([html]);
-  const buffer = Buffer.from(await htmlBlob.arrayBuffer());
-
-  writeFile(outputFilePath, buffer, () =>
-    console.log(`\nReport: ${URL.pathToFileURL(outFileFullDir)}`)
-  );
-};
 // eslint-disable-next-line complexity
 (async () => {
   const url = program.processedArgs?.[0] ?? program.opts().url;
@@ -366,7 +293,7 @@ const saveResultsAsHTML = async (
   const isSiteMap = sitemapUrl || filePath ? true : false;
 
   if (outDir) {
-    await saveCSVReports(path.resolve(outputDir), result);
+    await saveReports(path.resolve(outputDir), result);
     console.log('Reports created successfully!');
     process.exit(0);
   }
