+++ packages/cli/src/utils/index.ts
@@ -16,10 +16,11 @@
 export { fetchDictionary } from './fetchCookieDictionary';
 export { default as getUrlListFromArgs } from './getUrlListFromArgs';
 export { default as generateCSVFiles } from './generateCSVfiles';
-export { default as saveCSVReports } from './saveCSVReports';
+export { default as saveReports } from './saveReports';
 export { default as getUrlsFromSitemap } from './getUrlListFromArgs';
 export { default as askUserInput } from './askUserInput';
 export { default as generatePrefix } from './generatePrefix';
 export { default as getCurrentDateAndTime } from './getCurrentDateAndTime';
 export { default as getOutputFilePath } from './getOutputFilePath';
+export { default as saveResultAsHTML } from './saveResultAsHTML';
 export * from './validators';
